created components folder and components
