+++ components/added-exercises/index.js
@@ -1,0 +1,10 @@
+export default function AddedExercises({ addedExercise }) {
+  console.log("Added Exercise: ", addedExercise);
+  return (
+    <ul>
+      {addedExercise.map((exercise) => (
+        <li key={exercise}>{exercise}</li>
+      ))}
+    </ul>
+  );
+}
